[SYCL] Construct context from first device in unit tests
This is because multiple devices are not supported, and so when
constructing a context from a platform it tries to use all the devices
in the platform.

Using the first device should be fine because these tests should be run
with a device filter anyway. So this will only be an issue if there is
multiple devices of the same type, in which case using the first one for
testing should be fine.