[SYCL] Removed unused queues from property test
This patch removes the construction of unused queue instances from some
of the compile time property tests.

The queue doesn't seem to be needed there, and additionally the
`properties_get_property` test actually gets run, but the in-tree lit
tests are not supposed to require SYCL devices, so creating the queue
may cause issues on some setups.